LayoutTests/media/media-source/media-source-real-audio-append-buffer-expected.txt


EVENT(sourceopen)
EXPECTED (sourceBuffer.updating == 'true') OK
EVENT(updateend)
EXPECTED (sourceBuffer.updating == 'false') OK
EXPECTED (sourceBuffer.updating == 'true') OK
EVENT(updateend)
EXPECTED (sourceBuffer.updating == 'false') OK
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '4') OK
EXPECTED (sourceBuffer.audioTracks.length == '1') OK
EXPECTED (source.activeSourceBuffers.length == '1') OK
END OF TEST
